requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
appdirs=1.4.4=pypi_0
asgiref=3.3.1=pyhd3eb1b0_0
astroid=2.5.1=pypi_0
autopep8=1.5.5=pypi_0
ca-certificates=2021.1.19=haa95532_1
certifi=2020.12.5=py39haa95532_0
colorama=0.4.4=pypi_0
distlib=0.3.1=pypi_0
django=3.1.7=pyhd3eb1b0_0
djangorestframework=3.12.2=pypi_0
faker=7.0.1=pypi_0
filelock=3.0.12=pypi_0
isort=5.7.0=pypi_0
krb5=1.17.1=hc04afaa_0
lazy-object-proxy=1.5.2=pypi_0
libpq=12.2=h3235a2c_0
mccabe=0.6.1=pypi_0
openssl=1.1.1k=h2bbff1b_0
pip=21.0.1=py39haa95532_0
psycopg2=2.8.6=py39hcd4344a_1
pycodestyle=2.6.0=pypi_0
pylint=2.7.2=pypi_0
python=3.9.2=h6244533_0
python-dateutil=2.8.1=pypi_0
pytz=2021.1=pyhd3eb1b0_0
setuptools=52.0.0=py39haa95532_0
six=1.15.0=pypi_0
sqlite=3.35.3=h2bbff1b_0
sqlparse=0.4.1=py_0
text-unidecode=1.3=pypi_0
tk=8.6.10=he774522_0
toml=0.10.2=pypi_0
tzdata=2020f=h52ac0ba_0
vc=14.2=h21ff451_1
virtualenv=20.4.2=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.36.2=pyhd3eb1b0_0
wincertstore=0.2=py39h2bbff1b_0
wrapt=1.12.1=pypi_0
zlib=1.2.11=h62dcd97_4
